third_party/WebKit/LayoutTests/http/tests/security/cross-frame-access-frameelement.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFrames();
}
</script>
</head>
<body>
<p>This test checks if frameElement is accessible from the same or cross origin iframe.</p>
<iframe src="resources/cross-frame-access-frameelement-from-iframe.html"></iframe>
<iframe src="http://127.0.0.1:8080/security/resources/cross-frame-access-frameelement-from-iframe.html"></iframe>
</body>
</html>
