third_party/blink/web_tests/platform/generic/http/tests/security/cross-frame-access-frameelement-expected.txt

This test checks if frameElement is accessible from the same or cross origin iframe.

 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
[object HTMLIFrameElement]

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
null
